fix walk-in page build
